Error when delete of deleted model #2

+++ action/adapter/Delete.php
@@ -19,23 +19,28 @@ class Delete extends Adapter
     protected function _run()
     {
         $model = $this->getModel();
-        $isFailedDelete = false;
-        try {
-            $model->delete();
-        } catch (IntegrityException $e) {
-            $isFailedDelete = true;
-            $response = \yii::$app->response->redirect(\Yii::$app->request->referrer);
-            $flashes = ['kv-detail-warning' => $this->translate('Record') . ' #' . $model->primaryKey . ' ' . $this->translate('cannot be deleted because it has associated entries that cannot be deleted') . '. ' . $this->translate('Error content') . ': "' . $e->getMessage() . '".'];
+        if (!$model) {
+            $flashes = [
+                'kv-detail-warning' => $this->translate('You are trying to delete an already deleted entry.'),
+            ];
+        } else {
+            $isFailedDelete = false;
+            try {
+                $model->delete();
+            } catch (IntegrityException $e) {
+                $isFailedDelete = true;
+                $flashes = ['kv-detail-warning' => $this->translate('Record') . ' #' . $model->primaryKey . ' ' . $this->translate('cannot be deleted because it has associated entries that cannot be deleted') . '. ' . $this->translate('Error content') . ': "' . $e->getMessage() . '".'];
+            }
+
+            if (!$isFailedDelete) {
+                $flashes = ['kv-detail-success' => $this->translate('Record') . ' #' . $model->primaryKey . ' ' . $this->translate('is successfully') . ' ' . $this->translate('deleted')];
+            }
         }
 
-        if (!$isFailedDelete) {
-            if ($this->isRedirect) {
-                $response = \yii::$app->response->redirect(\Yii::$app->request->referrer);
-                $flashes = ['kv-detail-success' => $this->translate('Record') . ' #' . $model->primaryKey . ' ' . $this->translate('is successfully') . ' ' . $this->translate('deleted')];
-            } else {
-                $flashes = [];
-                $response = '';
-            }
+        if ($this->isRedirect) {
+            $response = \yii::$app->response->redirect(\Yii::$app->request->referrer);
+        } else {
+            $response = '';
         }
 
         $response = $this->getResponse([
